chore: move refresh to authentication
