--- packages/nestjs-auth-refresh/src/interfaces/auth-refresh-settings.interface.ts
@@ -1,4 +1,0 @@
-import { JwtStrategyOptionsInterface } from '@concepta/nestjs-authentication';
-
-export interface AuthRefreshSettingsInterface
-  extends JwtStrategyOptionsInterface {}
