+++ packages/nestjs-authentication/src/refresh/config/auth-refresh-default.config.ts
@@ -1,5 +1,5 @@
 import { registerAs } from '@nestjs/config';
-import { ExtractJwt } from '@concepta/nestjs-authentication';
+import { ExtractJwt } from '../../jwt/index';
 import { AUTH_REFRESH_MODULE_DEFAULT_SETTINGS_TOKEN } from '../auth-refresh.constants';
 import { AuthRefreshSettingsInterface } from '../interfaces/auth-refresh-settings.interface';
