+++ packages/nestjs-authentication/src/refresh/exceptions/auth-refresh-unauthorized.exception.ts
