+++ packages/nestjs-authentication/src/refresh/interfaces/auth-refresh-settings.interface.ts
@@ -1,0 +1,4 @@
+import { JwtStrategyOptionsInterface } from '../../jwt/interfaces/jwt-strategy-options.interface';
+
+export interface AuthRefreshSettingsInterface
+  extends JwtStrategyOptionsInterface {}
