+++ packages/nestjs-samples/tsconfig.json
@@ -11,10 +11,5 @@
   },
   "include": [
     "src/**/*.ts"
-  ],
-  "references": [
-    {
-      "path": "../nestjs-auth-refresh"
-    }
   ]
 }
